Avoid frequent data type casting (float/long) for rwi score refactor to using long in URIMetadataNode too (and related call parameters) As remote rwi score's are not used (since v1.83) skip reading float-score , but keep in toString() for communication with older versions.

+++ source/net/yacy/search/index/Fulltext.java
@@ -285,22 +285,33 @@ public final class Fulltext {
         getDefaultConnector().commit(softCommit);
         if (this.writeWebgraph) getWebgraphConnector().commit(softCommit);
     }
-    
+
+    /**
+     * Loads the meta data stored in the embedded solr index for the url referenced
+     * by the WordReference.
+     * The WordReference and the YaCy score (element.weight()) become part (and
+     * are accessible) of the returned document.
+     * If the no document with url.hash = solrdocument.id is found in the embedded
+     * Solr index null is return.
+     * 
+     * @param element rwi wordreference
+     * @return URIMetadataNode (solrdocument) with all fields stored in embedded solr index
+     */
     public URIMetadataNode getMetadata(final WeakPriorityBlockingQueue.Element<WordReferenceVars> element) {
         if (element == null) return null;
         WordReferenceVars wre = element.getElement();        
         if (wre == null) return null; // all time was already wasted in takeRWI to get another element
-        float score = element.getWeight();
+        long score = element.getWeight();
         URIMetadataNode node = getMetadata(wre.urlhash(), wre, score);
         return node;
     }
 
     public URIMetadataNode getMetadata(final byte[] urlHash) {
         if (urlHash == null) return null;
-        return getMetadata(urlHash, null, 0.0f);
+        return getMetadata(urlHash, null, 0L);
     }
     
-    private URIMetadataNode getMetadata(final byte[] urlHash, final WordReferenceVars wre, final float score) {
+    private URIMetadataNode getMetadata(final byte[] urlHash, final WordReferenceVars wre, final long score) {
         String u = ASCII.String(urlHash);
         
         // get the metadata from Solr
